Move around the selectors that create SamplesLikeTable objects
We'll need them in thread.js in the following commit.
